+++ tests/testthat/test-parse_hoboware_details.R
@@ -5,7 +5,8 @@ test_that("get_cond_details() works with single point cond. calibration", {
   files <- list.files(deployment, recursive = TRUE, full.names = TRUE)
   cond_path <- grep("_Cond_.*_Details.txt", files, value = TRUE)
   do_path <- grep("_DO_.*_Details.txt", files, value = TRUE)
-  expect_no_error(get_cond_details(cond_path))
+  expect_no_error(md <- get_cond_details(cond_path))
+  expect_no_error(md <- get_do_details(do_path))
   qc_deployment(deployment)
 
 })